+++ vortex-array/src/arrow/wrappers.rs
@@ -2,9 +2,9 @@ use arrow_buffer::{NullBuffer, OffsetBuffer, ScalarBuffer};
 use vortex_error::VortexResult;
 
 use crate::array::primitive::PrimitiveArray;
-use crate::array::Array;
 use crate::ptype::NativePType;
 use crate::validity::Validity;
+use crate::view::AsView;
 
 pub fn as_scalar_buffer<T: NativePType>(array: PrimitiveArray) -> ScalarBuffer<T> {
     assert_eq!(array.ptype(), T::PTYPE);
@@ -18,7 +18,7 @@ pub fn as_offset_buffer<T: NativePType>(array: PrimitiveArray) -> OffsetBuffer<T
 pub fn as_nulls(validity: Validity) -> VortexResult<Option<NullBuffer>> {
     match validity {
         Validity::Valid(_) => Ok(None),
-        Validity::Invalid(_) => Ok(Some(NullBuffer::new_null(validity.len()))),
+        Validity::Invalid(_) => Ok(Some(NullBuffer::new_null(validity.as_view().len()))),
         Validity::Array(_) => Ok(Some(NullBuffer::new(
             validity.to_bool_array().buffer().clone(),
         ))),
