+++ vortex-array/src/validity/mod.rs
@@ -1,7 +1,9 @@
 pub use array::*;
+pub use encoding::*;
 pub use owned::*;
 pub use view::*;
 
 mod array;
+mod encoding;
 mod owned;
 mod view;
