+++ vortex-fastlanes/src/bitpacking/mod.rs
@@ -141,7 +141,8 @@ impl Array for BitPackedArray {
         Self::try_new(
             self.encoded().slice(encoded_start, encoded_stop)?,
             self.validity()
-                .map(|v| v.slice(start, min(stop, self.len()))),
+                .map(|v| v.slice(start, min(stop, self.len())))
+                .transpose()?,
             self.patches()
                 .map(|p| p.slice(start, min(stop, self.len())))
                 .transpose()?,
