+++ vortex-ree/src/ree.rs
@@ -137,7 +137,10 @@ impl Array for REEArray {
         Ok(Self {
             ends: self.ends.slice(slice_begin, slice_end + 1)?,
             values: self.values.slice(slice_begin, slice_end + 1)?,
-            validity: self.validity().map(|v| v.slice(slice_begin, slice_end + 1)),
+            validity: self
+                .validity()
+                .map(|v| v.slice(slice_begin, slice_end + 1))
+                .transpose()?,
             offset: start,
             length: stop - start,
             stats: Arc::new(RwLock::new(StatsSet::new())),
